refactor: remove old link
