+++ packages/design/components/Typography/__test__/Link.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render } from '@testing-library/react'
-import Link from '../NewLink'
+import Link from '../Link'
 import { MemoryRouter } from 'react-router-dom'
 
 const LinkTestWrapper: React.FC = ({ children }) => {
