+++ packages/design/components/Typography/style/NewLink.less
@@ -1,12 +1,18 @@
 @import (reference) '../../../theme/base';
 
 .bgm-link {
-  font-size: 16px;
-  color: @blue-100;
-  font-weight: bold;
-  text-decoration: none;
+  &:link,
+  &:visited,
+  &:active {
+    color: @blue-100;
+    text-decoration: none;
+  }
 
   &:hover {
     text-decoration: underline;
+  }
+
+  &--bold {
+    font-weight: bold;
   }
 }
