+++ packages/website/src/pages/components/UserHome.module.less
@@ -8,7 +8,7 @@
   font-size: 24px;
 }
 
-.link {
-  color: @blue-100;
-  text-decoration: none;
-}
+// .link {
+//   color: @blue-100;
+//   text-decoration: none;
+// }
